Rename tagEditor tag-editor.ts
